Modify the header and move the search bar functionality into its own component.
